Added PopularPet & PetDetails Function.
